ciwiczenia 2013-12-08

+++ 7_funkcje/zadanie2.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
  
-long silnia(int);
+long long int silnia(int);
  
 int main()
 {
@@ -14,17 +14,17 @@ int main()
   else
   {
   
-  printf("%d! = %ld\n", liczba_naturalna, silnia(liczba_naturalna));
+  printf("%d! = %lli\n", liczba_naturalna, silnia(liczba_naturalna));
  
   }
   return 0;
   
 }
  
-long silnia(int x)
+long long int silnia(int x)
 {
   int a;
-  long out = 1;
+  long long int out = 1;
  
   for (a = 1; a <= x; a++)
     out = out * a;
